services/comments/requirements.txt
amqp==5.1.1
asgiref==3.7.2
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
kombu==5.3.1
Markdown==3.4.3
pytz==2023.3
sqlparse==0.4.4
vine==5.0.0
